refactor(model-postgres-json): Externalize query sorting compilation to PostgresJsonUtil.compileSort
